Extract TokenGenerators into its own file
